-upgrade to ESP-IDF v5
 -merged misc improvements from lightdesk
 -one-wire bus RMT implementation does not compile
 -port to new RMT driver rquired

+++ components/dev_pwm/include/dev_pwm/cmd_random.hpp
@@ -44,7 +44,7 @@ private:
   };
 
 private:
-  static size_t availablePrimes();
+  static uint32_t availablePrimes();
 
   static int32_t randomDirection();
   static uint32_t randomNum(uint32_t modulo);
